+++ src/main/java/com/jerry/ticketing/concert/application/ConcertQueryService.java
@@ -1,7 +1,7 @@
 package com.jerry.ticketing.concert.application;
 
 import com.jerry.ticketing.concert.domain.Concert;
-import com.jerry.ticketing.concert.infrastructure.repository.ConcertRepository;
+import com.jerry.ticketing.concert.domain.port.ConcertRepository;
 import com.jerry.ticketing.global.exception.BusinessException;
 import com.jerry.ticketing.global.exception.ConcertErrorCode;
 import lombok.RequiredArgsConstructor;
